split tests and modify build.yml
